This add Heartbeat as Uptime Monitors to Add Data UI
A new tab as is added to the Add Data UI page called Uptime. Clicking on Uptime there is currently one box with "Uptime Monitors" which guides through installing and configuring Heartbeat with 1 monitor and loading the dashboards.

If the setup is local Heartbeat by default will ping `http://localhost:9200`. The setup recommended for the Cloud setup is to ping `http://elastic.co`. It requires an additional step as it's a non default monitor but will allow to directly have data in the dashboards.